==> 2020_2 Fall/Data and Analysis/Data/CleaningData_OnlineExperiment/Jupyter Notebooks (Old)/Step3.ipynb <==
import numpy as np
import pandas as pd
from os import listdir
from os.path import isfile, join
import matplotlib.pyplot as plt
import warnings
import seaborn as sns
sns.set_style("darkgrid")
warnings.filterwarnings('ignore')
oloc = open("../Output_Location.txt", "r").read()
folder = ''
df = pd.read_csv(oloc+'Task1_step2.csv')
temp2 = pd.read_csv(oloc+'Task2.csv')
temp3 = pd.read_csv(oloc+'Task3.csv')

for key,value in df.iterrows():
    if value[1] == 0:
        temp = df.iloc[key, 8:17].to_list()
        df.iloc[key, 8:17] = [int(x) ^ 1 if x == 1.0 or x == 0.0 else np.nan for x in temp]
df.to_csv(oloc+'Task1_step3.csv', index=False)

folder = ''
df = pd.read_csv(oloc+'Task2_step2.csv')
for key,value in df.iterrows():
    if value[1] == 0:
        temp = df.iloc[key, 9:69].to_list()
        df.iloc[key, 9:69] = [int(x) ^ 1 if x == 1.0 or x == 0.0 else np.nan for x in temp]
df.to_csv(oloc+'Task2_step3.csv', index=False)

folder = ''
df = pd.read_csv(oloc+'Task3_step2.csv')
for key,value in df.iterrows():
    if value[1] == 0:
        temp = df.iloc[key, 9:69].to_list()
        df.iloc[key, 9:69] = [int(x) ^ 1 if x == 1.0 or x == 0.0 else np.nan for x in temp]
df.to_csv(oloc+'Task3_step3.csv', index=False)
